Moved psbt-wide prevout extraction/validation into application pkg.
