+++ pkg/arkade/psbt_fields.go
@@ -2,7 +2,6 @@ package arkade
 
 import (
 	"bytes"
-	"fmt"
 
 	"github.com/arkade-os/arkd/pkg/ark-lib/txutils"
 	"github.com/btcsuite/btcd/btcutil/psbt"
@@ -13,43 +12,6 @@ var (
 	ArkFieldPrevoutTx                                       = []byte("prevarktx")
 	PrevoutTxField    txutils.ArkPsbtFieldCoder[wire.MsgTx] = arkPsbtFieldCoderPrevoutTx{}
 )
-
-func PrevoutTxsFromPSBT(ptx *psbt.Packet) (map[int]*wire.MsgTx, error) {
-	if len(ptx.Inputs) != len(ptx.UnsignedTx.TxIn) {
-		return nil, fmt.Errorf("malformed psbt")
-	}
-
-	prevoutTxs := make(map[int]*wire.MsgTx)
-
-	for inputIndex := range ptx.Inputs {
-		fields, err := txutils.GetArkPsbtFields(ptx, inputIndex, PrevoutTxField)
-		if err != nil {
-			return nil, fmt.Errorf("failed to decode prevout tx for input %d: %w", inputIndex, err)
-		}
-
-		if len(fields) == 0 {
-			continue
-		}
-		if len(fields) > 1 {
-			return nil, fmt.Errorf("multiple prevout tx fields found for input %d", inputIndex)
-		}
-
-		prevTx := fields[0]
-		expectedHash := ptx.UnsignedTx.TxIn[inputIndex].PreviousOutPoint.Hash
-		actualHash := prevTx.TxHash()
-		if actualHash != expectedHash {
-			return nil, fmt.Errorf(
-				"prevout tx hash mismatch for input %d: got %s, expected %s",
-				inputIndex, actualHash, expectedHash,
-			)
-		}
-
-		prevTxCopy := prevTx
-		prevoutTxs[inputIndex] = &prevTxCopy
-	}
-
-	return prevoutTxs, nil
-}
 
 type arkPsbtFieldCoderPrevoutTx struct{}
